Move tray icon implementations into submodules with no other changes
